[Protein graphs] refactor to use Paths
