+++ graphein/protein/config.py
@@ -4,46 +4,51 @@
 # License: MIT
 # Project Website: https://github.com/a-r-j/graphein
 # Code Repository: https://github.com/a-r-j/graphein
-
+from __future__ import annotations
 from typing import Callable, List, Optional, Union
 
 from pydantic import BaseModel
 
-from graphein.features.amino_acid import meiler_embedding
-from graphein.features.edges.intramolecular import peptide_bonds
+from graphein.protein.features.nodes.amino_acid import meiler_embedding
+from graphein.protein.edges.intramolecular import peptide_bonds
 from pathlib import Path
+
+
+class GetContactsConfig(BaseModel):
+    get_contacts_path: Path = Path("/Users/arianjamasb/github/getcontacts/").resolve()
+    contacts_dir: Path = Path("../examples/contacts/").resolve()
+    pdb_dir: Path = Path("../examples/pdbs/").resolve()
+    granularity: str = "CA"
 
 
 class ProteinGraphConfig(BaseModel):
     granularity: str = "CA"
     keep_hets: bool = False
     insertions: bool = False
-    ### TODO: I suggest refactoring this out into a GetContactsConfig object.
-    # Also suggest to avoid hard-coding paths if possible!
-    get_contacts_path: Optional[Path] = None
-    pdb_dir: Path = "../examples/pdbs/"
-    contacts_dir: str = "../examples/contacts/"
-    ### END TODO
+    pdb_dir: Path = Path("../examples/pdbs/")  # Also suggest to avoid hard-coding paths if possible!
     verbose: bool = True
     exclude_waters: bool = True
     covalent_bonds: bool = True
     include_ss: bool = True
     include_ligand: bool = False
-    intramolecular_interactions: Optional[List[str]] = None  # Todo rm
     graph_constructor: Optional[str] = None
-    edge_distance_cutoff: Optional[float] = None
     verbose: bool = True
     deprotonate: bool = False
-    remove_string_labels: bool = False
     long_interaction_threshold: Optional[int] = None
+    data_frame_processing_functions: Optional[List[Callable]] = None
     edge_construction_functions: List[Union[Callable, str]] = [peptide_bonds]
     node_metadata_functions: Optional[List[Union[Callable, str]]] = [
         meiler_embedding
     ]
     edge_metadata_functions: Optional[List[Union[Callable, str]]] = None
     graph_metadata_functions: Optional[List[Callable]] = None
+    get_contacts_config: Optional[GetContactsConfig] = GetContactsConfig()
+
+    class Config:
+        arbitrary_types_allowed: bool = True
 
 
 class ProteinMeshConfig(BaseModel):
     pymol_command_line_options: Optional[str] = "-cKq"
     pymol_commands: Optional[List[str]] = ["show surface"]
+
