+++ graphein/protein/edges/intramolecular.py
