+++ graphein/protein/features/data/amino_acid_properties.csv
