+++ graphein/protein/utils.py
@@ -5,7 +5,7 @@
 # Project Website: https://github.com/a-r-j/graphein
 # Code Repository: https://github.com/a-r-j/graphein
 import os
-
+from pathlib import Path
 from Bio.PDB import PDBList
 
 
@@ -18,7 +18,7 @@ def download_pdb(config, pdb_code: str) -> str:
     :return: # todo impl return
     """
     if not config.pdb_dir:
-        config.pdb_dir = "/tmp/"
+        config.pdb_dir = Path("/tmp/")
 
     # Initialise class and download pdb file
     pdbl = PDBList()
@@ -27,13 +27,13 @@ def download_pdb(config, pdb_code: str) -> str:
     )
     # Rename file to .pdb from .ent
     os.rename(
-        config.pdb_dir + "pdb" + pdb_code + ".ent",
-        config.pdb_dir + pdb_code + ".pdb",
+        config.pdb_dir / ("pdb" + pdb_code + ".ent"),
+        config.pdb_dir / (pdb_code + ".pdb"),
     )
     # Assert file has been downloaded
     assert any(pdb_code in s for s in os.listdir(config.pdb_dir))
     print(f"Downloaded PDB file for: {pdb_code}")
-    return config.pdb_dir + pdb_code + ".pdb"
+    return config.pdb_dir / (pdb_code + ".pdb")
 
 
 def get_protein_name_from_filename(pdb_path: str) -> str:
